WIP Support HttpClient based scenarios
When JPT detects a regression and the regression is on a backend side, then
it's handy to test only the backend side, to avoid the browser's noise.
I had this problem twice and two persons asked me about the feature. As there was
no dedicated API, I've been creating a custom scenario, and translate web driver
to HttpClient inside of a custom action.

This approach has two disadvantages:

 - It requires expert-level knowledge about JPT
 - Each VU needs to run a real browser, even if it's not used. It limits the

The goal of this change is to expose API for lightweight HttpClient based scenarios.