Unignore `EntryPointIT`
Unblock the main API integration test.
Here's the [original rationale] for ignoring it.
In the meantime there was some progress from `testcontainers` to [reduce the class copy-paste].
This allows us to work with `docker-java-api`, which no longer clashes.
However, they're still shading `docker-java-core`, so we gotta depend on their copy for now.
Currently their copy is compatible with the rest of our code.
It's still messy, but we can stop depending on them later.

[original rationale]: https://github.com/atlassian/virtual-users/commit/29c29fc6d66e4376b0b8ac3c127503211a69cc03
[reduce the class copy-paste] : https://github.com/testcontainers/testcontainers-java/pull/2882