JPERF-706: Fix Kotlin compilation problem
### The error
Fix error:
```
Could not perform incremental compilation: Could not connect to Kotlin compile daemon
Could not connect to kotlin daemon. Using fallback strategy.

warning: unable to find kotlin-stdlib.jar in the Kotlin home directory. Pass either '-no-stdlib' to prevent adding it to the classpath, or the correct '-kotlin-home'
warning: unable to find kotlin-script-runtime.jar in the Kotlin home directory. Pass either '-no-stdlib' to prevent adding it to the classpath, or the correct '-kotlin-home'
warning: unable to find kotlin-reflect.jar in the Kotlin home directory. Pass either '-no-reflect' or '-no-stdlib' to prevent adding it to the classpath, or the correct '-kotlin-home'
error: source file or directory not found: @/Users/mkwidzinski/Work/jpt/virtual-users/build/20201110_2299434203285249803.compiler.options
```

The `-no-stdlib` flag does not fix it.

### Bumping Kotlin compiler classpath
The kotlin compiler has its own classpath, which is included in our dependency locks.
So we could bump it so it's compatible with the bumped Kotlin compiler.
But bumping it comes with some complications.

### Bumping all classpaths
If we bump it for all classpaths, we affect the production and test code.
Kotlin bump from 1.2 to 1.3 is a breaking change. According to [Kotlin evolution], a "feature" release can break code:
> Feature releases often add new features and may remove or change previously deprecated ones.

This would require a SemVer MAJOR bump of the `kotlin-stdlib-jdk8` dependency, which is breaking our POM contract.
It'd require a MAJOR release of `virtual-users`.

[Kotlin evolution]: https://kotlinlang.org/docs/reference/evolution/kotlin-evolution.html#feature-releases-and-incremental-releases

### Bumping compiler classpath only
It's possible, e.g.:
```
when {
    name.startsWith("kotlinCompiler") -> eachDependency {
        when (requested.group) {
            "org.jetbrains.kotlin" -> useVersion("1.4.10")
        }
    }
}
```
Selective bump `./gradlew dep --update-locks "org.jetbrains.kotlin:*"` fails with dependency conflicts.
So a `./gradlew dep --write-locks` would be necessary.
It can be done, but it grows the scope of the PR bit beyond "fix warnings".
Also, it's a complex thing to do the next time.
But we can make it simpler.

### Excluding compiler from locks
We can simply remove the dependency locking for the compiler.
There's not much value in keeping them, because "normal" dependencies don't affect it.
There are no transitive version conflicts to resolve.
This is the approach used in this commit. Should be less hassle going forward.