JPERF-706: Fix compilation compatibility warning
Fix warning:
```
w: Runtime JAR files in the classpath have the version 1.2, which is older than the API version 1.4.
Consider using the runtime of version 1.4, or pass '-api-version 1.2' explicitly to restrict the available APIs
to the runtime of version 1.2. You can also pass '-language-version 1.2' instead,
which will restrict not only the APIs to the specified version, but also the language features
w: /Users/mkwidzinski/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.2.70/kotlin-stdlib-common-1.2.70.jar:
Runtime JAR file has version 1.2 which is older than required for API version 1.4
```

Now we're getting warned about 1.2 being old:
```
w: Language version 1.2 is deprecated and its support will be removed in a future version of Kotlin
```

But that's a true warning. 1.2 is old and we gotta port to 1.3 or 1.4 at some point.
That requires a major release, as explained in 24cdab615dd4c9d7a58cfeb38c1df3585324de6d.